+++ utils/neo_config.cpp
@@ -1,7 +1,8 @@
 /* SPDX-License-Identifier: LGPL-2.1-or-later */
 /*
- * neo-config.cpp - Configuration helpers for IPA
  * Copyright 2025-2026 NXP
+ *
+ * Configuration helpers for IPA
  */
 
 #include "neo_config.h"
@@ -13,11 +14,11 @@ namespace libcamera::ipa::nxpneo {
 
 LOG_DEFINE_CATEGORY(NxpNeoUguzziConfig)
 
-const std::map<std::string, IPAModeType> IPAFileConfig::kIPAModeNameMap = {
-	{ "standard", IPAModeTypeStandard },
-	{ "hdr", IPAModeTypeHdrMerge },
-	{ "rgbIr", IPAModeTypeRgbIr },
-	{ "rgbIrDual", IPAModeTypeRgbIrDual },
+const std::map<std::string, IPAPipelineMode> IPAFileConfig::kPipelineModeNameMap = {
+	{ "standard", IPAPipelineMode::Standard },
+	{ "hdr", IPAPipelineMode::HdrMerge },
+	{ "rgbIr", IPAPipelineMode::RgbIr },
+	{ "rgbIrDual", IPAPipelineMode::RgbIrDual },
 };
 
 /**
@@ -160,19 +161,28 @@ int IPAFileConfig::parseSensorProfiles(const YamlObject &profiles,
 
 		const YamlObject &modeObj = profile["mode"];
 		const std::string modeName = modeObj.get<std::string>().value_or("");
-		auto iter = kIPAModeNameMap.find(modeName);
-		tuningInfo.mode = (iter != kIPAModeNameMap.end()) ? iter->second : kMode;
+		auto iter = kPipelineModeNameMap.find(modeName);
+		if (iter != kPipelineModeNameMap.end()) {
+			tuningInfo.mode = iter->second;
+		} else {
+			LOG(NxpNeoUguzziConfig, Warning)
+				<< "No pipeline mode found or valid for "
+				<< sensor << ": use standard mode by default.";
+			tuningInfo.mode = IPAPipelineMode::Standard;
+		}
 
 		const YamlObject &dtpObj = profile["dtp-file"];
 		tuningInfo.dtpFile = dtpObj.get<std::string>().value_or("");
 
+		static const std::vector<uint32_t> tuningIdDefault =
+			{ kTuningIdRgb, kTuningIdIr };
 		const YamlObject &tuningIdObj = profile["tuning-id"];
 		tuningInfo.tuningId =
-			tuningIdObj.get<uint16_t>().value_or(kTuningId);
+			tuningIdObj.getList<uint32_t>().value_or(tuningIdDefault);
 
 		const YamlObject &tuningModeObj = profile["tuning-mode"];
 		tuningInfo.tuningMode =
-			tuningModeObj.get<uint16_t>().value_or(kTuningMode);
+			tuningModeObj.get<uint32_t>().value_or(kTuningMode);
 
 		tuningInfo.resolution = resolution.value();
 
@@ -219,7 +229,7 @@ int IPAFileConfig::parseOverrideInAlign(const YamlObject &overrideInAlign)
  * \param[in] entity The name of the camera media device entity
  * \param[in] resolution The resolution of the camera stream
  * \param[in] bitDepth The bits per pixel of the camera stream
- * \param[in] mode The sensor stream mode
+ * \param[in] mode The pipeline mode type
  *
  * The tuningInfo structure contains information related to the tuning
  * of the sensor.
@@ -230,7 +240,7 @@ const TuningInfo *IPAFileConfig::tuningInfo(const std::string &model,
 					    const std::string &entity,
 					    Size resolution,
 					    unsigned int bitDepth,
-					    IPAModeType mode) const
+					    IPAPipelineMode mode) const
 {
 	/*
 	 * For the tuning info search, give priority to entity-based match over
@@ -255,14 +265,19 @@ const TuningInfo *IPAFileConfig::tuningInfo(const std::string &model,
 				});
 			if (iter_res != tuningInfos->end()) {
 				const TuningInfo *tuningInfo = &(*iter_res);
+				std::stringstream ssTuningId;
+				ssTuningId << "{";
+				for (auto id : tuningInfo->tuningId)
+					ssTuningId << " " << id;
+				ssTuningId << " }";
 				LOG(NxpNeoUguzziConfig, Debug)
 					<< "TuningInfo parsed for ["
 					<< entity << "; "
 					<< resolution << "; "
 					<< bitDepth << "bpp; mode:"
-					<< mode << "]: ["
+					<< static_cast<int>(mode) << "]: ["
 					<< tuningInfo->dtpFile << ", "
-					<< tuningInfo->tuningId << ", "
+					<< ssTuningId.str() << ", "
 					<< tuningInfo->tuningMode << "]";
 				return tuningInfo;
 			}
@@ -270,8 +285,9 @@ const TuningInfo *IPAFileConfig::tuningInfo(const std::string &model,
 	}
 	LOG(NxpNeoUguzziConfig, Error) << "No tuning Info found for ["
 				       << entity << "; "
-				       << resolution << "; " << bitDepth << "bpp; mode:"
-				       << mode << "]";
+				       << resolution << "; " << bitDepth
+				       << "bpp; mode:"
+				       << static_cast<int>(mode) << "]";
 	return nullptr;
 }