+++ utils/neo_config.h
@@ -1,7 +1,8 @@
 /* SPDX-License-Identifier: LGPL-2.1-or-later */
 /*
- * neo_config.h - Configuration helpers for IPA
  * Copyright 2025-2026 NXP
+ *
+ * Configuration helpers for IPA
  */
 
 #pragma once
@@ -22,10 +23,10 @@ namespace libcamera::ipa::nxpneo {
 struct TuningInfo {
 	Size resolution;
 	unsigned int bitDepth;
-	IPAModeType mode;
+	IPAPipelineMode mode;
 	std::string dtpFile;
-	unsigned int tuningId;
-	unsigned int tuningMode;
+	std::vector<uint32_t> tuningId;
+	uint32_t tuningMode;
 };
 
 using SensorMap = std::map<std::string, std::vector<TuningInfo>>;
@@ -41,7 +42,7 @@ public:
 				     const std::string &entity,
 				     Size resolution,
 				     unsigned int bitDepth,
-				     IPAModeType mode) const;
+				     IPAPipelineMode mode) const;
 	uint16_t socketPort(const std::string &model,
 			    const std::string &entity) const;
 	const std::optional<std::string> &sensorFilter() const { return sensorFilter_; }
@@ -50,9 +51,9 @@ public:
 private:
 	static constexpr uint16_t kSocketPort = 50000;
 	static constexpr unsigned int kBitDepth = 16;
-	static constexpr IPAModeType kMode = IPAModeTypeStandard;
-	static constexpr unsigned int kTuningId = 2010;
-	static constexpr unsigned int kTuningMode = 0;
+	static constexpr uint32_t kTuningIdRgb = 2010;
+	static constexpr uint32_t kTuningIdIr = 2011;
+	static constexpr uint32_t kTuningMode = 0;
 
 	int parseSensors(const YamlObject &sensors);
 	int parseSensorProfiles(const YamlObject &profiles,
@@ -64,8 +65,11 @@ private:
 	std::optional<std::string> sensorFilter_;
 	bool overrideInAlign_ = true;
 	SocketMap socketMap_;
-	/* Map between the sensor stream mode name and its associated IPA enum value. */
-	static const std::map<std::string, IPAModeType> kIPAModeNameMap;
+	/*
+	 * Map between the pipeline mode name and its associated IPA pipeline
+	 * mode enum value.
+	 */
+	static const std::map<std::string, IPAPipelineMode> kPipelineModeNameMap;
 };
 
 } /* namespace libcamera::ipa::nxpneo */